% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/atg_ex.R
\name{atg_ex}
\alias{atg_ex}
\title{Target group and filter out excluded antigen}
\usage{
atg_ex(df, HA = TRUE, atg_ex, group_name, grade_label = "")
}
\arguments{
\item{df}{Original dataset}

\item{HA}{Boolean, indicating whether used for HA analysis or HSDA analysis}

\item{atg_ex}{List of all excluded antigen in the original dataset}

\item{group_name}{String of group name}

\item{grade_label}{String of grade in 2 digits converted from group_name}
}
\value{
Dataset with only correct antigen and HA / HSDA of target group
}
\description{
Select correct grade / year level & HA / HSDA
and exclude unnecessary antigen
}
\examples{
df_g <- data.frame(
grade = c("06","06","07","08", "09", "10", "10", "11", "12"),
antigen = c("dtap", "dtap_p", "hpv_F", "hpv_M", "men", "hpv", "measle",
"td", "poilo"),
HSDA = c("Vancouver", "Richmond", "VCHA", "Richmond", "Vancouver",
"VCHA", "VCHA", "Vancouver", "Richmond")
)

df_y<- data.frame(
antigen = c("dtap", "dtap_p", "hpv_F", "hpv_M", "men", "hpv", "measle",
"td", "poilo"),
HSDA = c("Vancouver", "Richmond", "VCHA", "Richmond", "Vancouver",
"VCHA", "VCHA", "Vancouver", "Richmond")
)

atg_ex(df_y, TRUE, c("dtap", "dtap_p", "hpv_F", "hpv_M"), "7yr")
atg_ex(df_g, TRUE, c("dtap", "dtap_p", "hpv_F", "hpv_M"), "7gr","07")
atg_ex(df_g, FALSE, "td","12gr","12")

}
