% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/output_file.R
\name{output_file}
\alias{output_file}
\title{Path Generator}
\usage{
output_file(group_name, suffix, root_dir = "result")
}
\arguments{
\item{group_name}{String of group name from c("7yr","6gr","7gr","8gr","9gr",
"10gr","11gr","12gr")}

\item{suffix}{String of table type from c("coverage_final",
"reasons_for_Non_Imms")}

\item{root_dir}{String of root dirctory; defult: "result"}
}
\value{
Absolute path including the file name to the result dataset
}
\description{
Generate absoulte path to the output dataset
}
\examples{
output_file("7yr", "coverage_final")
output_file("12gr", "reasons_for_Non_Imms")
output_file("12gr", "reasons_for_Non_Imms","test_result_newcode")

}
