% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trans_cvg.R
\name{trans_cvg}
\alias{trans_cvg}
\title{Dataset Preparation}
\usage{
trans_cvg(group_name, grade_label = "", df = vch_df0)
}
\arguments{
\item{group_name}{A string selected from c("7yr","6gr","7gr","8gr","9gr",
"10gr","11gr","12gr")}

\item{grade_label}{String of grade in 2 digits converted from group_name}

\item{df}{Original dataset}
}
\value{
A list in length 9, including

vch_df_ha: dataset for HA Coverage analysis

vch_df_hsda: dataset for HDSA Coverage analysis

disease_tbmutate: string of a disease to be deeper explored

afternrows: number of the row to inserted the mutated disease

disease_rename: string of the new name of the variant disease

vch_hpv_ha: dataset for HA HPV related analysis

vch_hpv_hsda: dataset for HSDA related analysis

reason_df_ha: dataset for HA reason analysis

reason_df_hsda: dataset for HSDA reason analysis
}
\description{
Initial manipulate the dataset as per group for Coverage analysis
}
\examples{
vch_df0 <- data.frame(
antigen = c("dtap", "dtap_p", "hpv_F", "hpv_M", "men", "hpv", "measle",
"td", "poilo"),
HSDA = c("Vancouver", "Richmond", "VCHA", "Richmond", "Vancouver",
"VCHA", "VCHA", "Vancouver", "Richmond")
)
trans_cvg("7yr")

vch_df_g <- data.frame(
grade = c("06","06","07","08", "09", "10", "10", "11", "12"),
antigen = c("dtap", "dtap_p", "hpv_F", "hpv_M", "men", "hpv", "measle",
"td", "poilo"),
HSDA = c("Vancouver", "Richmond", "VCHA", "Richmond", "Vancouver",
"VCHA", "VCHA", "Vancouver", "Richmond")
)
trans_cvg("6gr", "06", vch_df_g)


}
