% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/var_assign.R
\name{var_assign}
\alias{var_assign}
\title{Title Select sheet, columns and diseases for Coverage&Reasons analysis}
\usage{
var_assign(group_name, cnames0, Disease0)
}
\arguments{
\item{group_name}{A string selected from
c("7yr","6gr","7gr","8gr","9gr","10gr","11gr","12gr")}

\item{cnames0}{A list of strings including all column names of
all the spreadsheets in the dataset}

\item{Disease0}{A list of strings including all disease, variation and notes
related to Coverage among all ages&grades in the dataset}
}
\value{
A list of assigned variables in length of 4, including
group_sheet, cnames, Disease_coverage and Disease_reason as per group

group_sheet: String of the sheet name in the dataset to be analyzed
cnames: Vector of the codable column names in selected sheet
Disease_coverage: Vector of readable diseases&variants&notes to be listed in the Coverage file
Disease_reason: Vector of readable diseases to be listed in the NonImmReason file
}
\description{
Title Select sheet, columns and diseases for Coverage&Reasons analysis
}
\examples{
cols = c("antigen", "HSDA", "students", "refusal_all.n",
"refusal_all.percent", "no_imm_record.n", "no_imm_record.percent",
"utd_all.n", "utd_all.percent", "utd.n", "utd.percent", "immunity.n",
"immunity.percent", "partial_contra.n", "partial_contra.percent",
"partial_refusal.n", "partial_refusal.percent", "partial_other.n",
"partial_other.percent", "unimm_contra.n", "unimm_contra.percent",
"unimm_refusal.n", "unimm_refusal.percent", "unimm_other.n",
"unimm_other.percent")

dis0 = c("1UptoDate","D_T_aP", "D_T_aP_IPV", "HepB", "HPV",
"HPV Series Initiation but not Completion", "HPV-Female", "HPV-Male",
"Measles", "Measles - 1 Dose Only", "Meningo", "Mumps", "Pertussis", "Polio",
"Rubella", "Td", "Varicella", "xNo Information", "yRefusalstoAll&NoImms")

var_assign("7gr", cols, dis0)
var_assign("7yr", cols, dis0)
var_assign("12gr", cols, dis0)

}
